requirements.txt
streamlit
deepface
opencv-python-headless
Pillow
numpy
